# Some throughput improvements for Simple Meters with decreasing count of created objects during measurement process