--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,5 @@
 build/
 .gradle/
 gradle.properties
+*.iml
+.idea