--- a/micrometer-core/build.gradle
+++ b/micrometer-core/build.gradle
@@ -47,8 +47,5 @@ dependencies {
 
 jmh {
     jmhVersion = '1.18'
-    warmupIterations = 2
-    iterations = 5
-    fork = 0
     duplicateClassesStrategy = 'warn'
 }