--- a/micrometer-core/src/jmh/java/io/micrometer/core/benchmark/QuantilesBenchmark.java
+++ b/micrometer-core/src/jmh/java/io/micrometer/core/benchmark/QuantilesBenchmark.java
@@ -27,6 +27,9 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 import java.util.concurrent.TimeUnit;
 
+@Warmup(iterations = 2)
+@Fork(value = 0)
+@Measurement(iterations = 5)
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
 @State(Scope.Benchmark)