--- a/micrometer-core/src/main/java/io/micrometer/core/instrument/simple/SimpleDistributionSummary.java
+++ b/micrometer-core/src/main/java/io/micrometer/core/instrument/simple/SimpleDistributionSummary.java
@@ -15,26 +15,35 @@
  */
 package io.micrometer.core.instrument.simple;
 
+import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Measurement;
 import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.util.MeterId;
-import io.micrometer.core.instrument.DistributionSummary;
 
 import java.util.Arrays;
 import java.util.concurrent.atomic.DoubleAdder;
 import java.util.concurrent.atomic.LongAdder;
 
 public class SimpleDistributionSummary extends AbstractSimpleMeter implements DistributionSummary {
+    private static final Tag STAT_COUNT_TAG = Tag.of("statistic", "count");
+    private static final Tag STAT_AMOUNT_TAG = Tag.of("statistic", "amount");
+    private static final Tag TYPE_TAG = SimpleUtils.typeTag(Type.DistributionSummary);
+
+    private final MeterId countId;
+    private final MeterId amountId;
+
     private LongAdder count = new LongAdder();
     private DoubleAdder amount = new DoubleAdder();
 
     public SimpleDistributionSummary(MeterId id) {
         super(id);
+        this.countId = id.withTags(TYPE_TAG, STAT_COUNT_TAG);
+        this.amountId = id.withTags(TYPE_TAG, STAT_AMOUNT_TAG);
     }
 
     @Override
     public void record(double amount) {
-        if(amount >= 0) {
+        if (amount >= 0) {
             count.increment();
             this.amount.add(amount);
         }
@@ -53,8 +62,7 @@ public class SimpleDistributionSummary extends AbstractSimpleMeter implements Di
     @Override
     public Iterable<Measurement> measure() {
         return Arrays.asList(
-            id.withTags(SimpleUtils.typeTag(getType()), Tag.of("statistic", "count")).measurement(count()),
-            id.withTags(SimpleUtils.typeTag(getType()), Tag.of("statistic", "amount")).measurement(totalAmount())
-        );
+                countId.measurement(count()),
+                amountId.measurement(totalAmount()));
     }
 }