--- a/micrometer-core/src/main/java/io/micrometer/core/instrument/simple/SimpleGauge.java
+++ b/micrometer-core/src/main/java/io/micrometer/core/instrument/simple/SimpleGauge.java
@@ -17,6 +17,7 @@ package io.micrometer.core.instrument.simple;
 
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.Measurement;
+import io.micrometer.core.instrument.Tag;
 import io.micrometer.core.instrument.util.MeterId;
 
 import java.lang.ref.WeakReference;
@@ -24,13 +25,16 @@ import java.util.Collections;
 import java.util.function.ToDoubleFunction;
 
 public class SimpleGauge<T> extends AbstractSimpleMeter implements Gauge {
+    private static final Tag TYPE_TAG = SimpleUtils.typeTag(Type.Gauge);
     private final WeakReference<T> ref;
     private final ToDoubleFunction<T> value;
+    private final MeterId gaugeId;
 
     public SimpleGauge(MeterId id, T obj, ToDoubleFunction<T> value) {
         super(id);
         this.ref = new WeakReference<>(obj);
         this.value = value;
+        this.gaugeId = id.withTags(TYPE_TAG);
     }
 
     @Override
@@ -40,6 +44,6 @@ public class SimpleGauge<T> extends AbstractSimpleMeter implements Gauge {
 
     @Override
     public Iterable<Measurement> measure() {
-        return Collections.singletonList(id.withTags(SimpleUtils.typeTag(getType())).measurement(value()));
+        return Collections.singletonList(gaugeId.measurement(value()));
     }
 }