This can be useful for all backends in #48 
```
e.g: on MBP 2014  (with Xmx8m)
Before fix
SimpleMeasureBenchmark.timerMeasure                  thrpt   30  546,673 ± 5,144  ops/ms
SimpleMeasureBenchmark.longTaskTimerMeasure          thrpt   30  480,073 ± 10,585  ops/ms
SimpleMeasureBenchmark.counterMeasure                thrpt   30  1158,556 ± 13,378  ops/ms
SimpleMeasureBenchmark.gaugeMeasure                  thrpt   30  1119,509 ± 14,784  ops/ms
SimpleMeasureBenchmark.distributionSummaryMeasure    thrpt   30  514,463 ± 5,484  ops/ms

After fix
SimpleMeasureBenchmark.timerMeasure                  thrpt   30  2088,021 ± 14,168  ops/ms
SimpleMeasureBenchmark.longTaskTimerMeasure          thrpt   30  1665,487 ± 17,270  ops/ms
SimpleMeasureBenchmark.counterMeasure                thrpt   30  5129,027 ± 48,356  ops/ms
SimpleMeasureBenchmark.gaugeMeasure                  thrpt   30  4904,147 ± 55,073  ops/ms
SimpleMeasureBenchmark.distributionSummaryMeasure    thrpt   30  2040,290 ± 15,672  ops/ms
```