@checketts 2017-07-22
Looks like you'll want to rebase to remove merge conflicts. 4x improvement is pretty impressive!